requirements.txt
django
python-decouple
djangorestframework
drf-spectacular
